requirements.txt
anyio==4.9.0
APScheduler==3.11.0
beautifulsoup4==4.13.5
bidict==0.23.1
blinker==1.9.0
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
DateTime==5.5
deep-translator==1.11.4
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
Flask-SocketIO==5.5.1
Flask-WTF==1.2.2
google_search_results==2.4.2
greenlet==3.2.4
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
imbalanced-learn==0.14.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.3
playwright==1.55.0
psutil==7.1.0
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
python-telegram-bot==22.3
pytz==2025.2
requests==2.32.4
rfc3986==1.5.0
schedule==1.2.2
scheduler==0.8.8
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
threadpoolctl==3.6.0
typeguard==4.4.4
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
xlrd==2.0.2
xlwt==1.3.0
zope.interface==7.2
